:white_check_mark: SteadyQueryRepositoryImplTest V2로 변경
